# docs/api/dev/orders-service.openapi.yaml
openapi: 3.0.3
info:
  title: Orders Service API
  version: 1.0.0
  description: Orders microservice for DashRoute
servers:
  - url: http://localhost:4002
    description: Local development server

paths:
  /orders:
    post:
      summary: Create a new order
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderDto'
      responses:
        '201':
          description: Order created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDto'
    get:
      summary: Get orders (Paginated)
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          description: Max number of orders to return (default 10)
          schema:
            type: integer
            default: 10
            maximum: 50
        - name: cursor
          in: query
          description: Pagination cursor string
          schema:
            type: string
        - name: status
          in: query
          description: Filter by status
          schema:
            type: string
            enum: [PENDING, ASSIGNED, ACCEPTED, IN_TRANSIT, DELIVERED, CANCELLED]
        - name: customerId
          in: query
          description: Filter by customer ID (ADMIN only)
          schema:
            type: string
            format: uuid
        - name: courierId
          in: query
          description: Filter by courier ID (ADMIN only)
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Paginated list of orders
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/OrderDto'
                  nextCursor:
                    type: string
                    nullable: true
                  hasNextPage:
                    type: boolean
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  /orders/{id}:
    get:
      summary: Get a specific order by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Order ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Order details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDto'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          description: Order not found

  /orders/{id}/accept:
    patch:
      summary: Accept an assigned order
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Order ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Order accepted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDto'
        '400':
          description: Invalid state transition
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Forbidden (Courier not assigned to order)
        '404':
          description: Order not found

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    LocationDto:
      type: object
      required:
        - lat
        - lng
      properties:
        lat:
          type: number
          example: 40.7128
        lng:
          type: number
          example: -74.0060

    CreateOrderDto:
      type: object
      required:
        - pickupLocation
        - dropoffLocation
      properties:
        pickupLocation:
          $ref: '#/components/schemas/LocationDto'
        dropoffLocation:
          $ref: '#/components/schemas/LocationDto'

    OrderDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        customerId:
          type: string
          format: uuid
        courierId:
          type: string
          format: uuid
          nullable: true
        status:
          type: string
          enum: [PENDING, ASSIGNED, ACCEPTED, IN_TRANSIT, DELIVERED, CANCELLED]
        pickupLocation:
          $ref: '#/components/schemas/LocationDto'
        dropoffLocation:
          $ref: '#/components/schemas/LocationDto'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

  responses:
    UnauthorizedError:
      description: Unauthorized. Missing or invalid token.
      content:
        application/json:
          schema:
            type: object
            properties:
              statusCode:
                type: integer
                example: 401
              message:
                type: string
                example: 'Authentication token missing'